refactor(explore_page): separate widgets, models, and services into dedicated folders
