+++ lib/models/blog_item.dart
@@ -1,0 +1,9 @@
+class BlogItem {
+  final String title;
+  final String thumbnailUrl;
+
+  BlogItem({
+    required this.title,
+    required this.thumbnailUrl,
+  });
+}
