+++ lib/models/video_item.dart
@@ -1,0 +1,11 @@
+class VideoItem {
+  final String title;
+  final String thumbnailUrl;
+  final String videoId;
+
+  VideoItem({
+    required this.title,
+    required this.thumbnailUrl,
+    required this.videoId,
+  });
+}
