+++ lib/widgets/horizontal_card_list.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+
+class HorizontalCardList extends StatelessWidget {
+  final String title;
+  final List<Widget> cards;
+
+  const HorizontalCardList({
+    super.key,
+    required this.title,
+    required this.cards,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 16.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16.0),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(title, style: const TextStyle(fontSize: 14, fontWeight: FontWeight.w700)),
+                const Text(
+                  'See more',
+                  style: TextStyle(color: Color(0xFF006FFD), fontSize: 12, fontWeight: FontWeight.w600),
+                ),
+              ],
+            ),
+          ),
+          const SizedBox(height: 8),
+          SizedBox(
+            height: 220,
+            child: ListView.separated(
+              scrollDirection: Axis.horizontal,
+              padding: const EdgeInsets.symmetric(horizontal: 16),
+              itemCount: cards.length,
+              separatorBuilder: (_, __) => const SizedBox(width: 12),
+              itemBuilder: (context, index) => cards[index],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
